latest

+++ Tech-Store Managment'/src/view_models/PurchaseStock.java
@@ -1,5 +1,7 @@
 package view_models;
 
+import java.util.Optional;
+
 import controllers.StockController;
 import interfaces.ViewException;
 import javafx.collections.FXCollections;
@@ -8,6 +10,7 @@ import javafx.geometry.Insets;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.Label;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
@@ -15,19 +18,26 @@ import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.layout.BorderPane;
 import javafx.scene.layout.GridPane;
 import javafx.scene.layout.VBox;
+import javafx.stage.Stage;
 import models.StockItem;
 
 public class PurchaseStock extends BorderPane {
 
 	private StockView leftSide = new StockView();
-	private QuantityPlusMinus quantity = new QuantityPlusMinus(1);
-	private GridPane purchaseSide = new GridPane();
 	
 	
 	public PurchaseStock() {
-		setPadding(new Insets(30));
 		setCenter(leftSide);
-		initPurchaseSide();
+	}
+	
+	
+	
+	public void purchaseItem() {
+			try {
+				leftSide.purchaseItem();
+			} catch (ViewException e) {
+				e.showAlert();
+			}
 	}
 	
 	public void delete() {
@@ -42,58 +52,41 @@ public class PurchaseStock extends BorderPane {
 		leftSide.refresh();
 	}
 	
-	private void initPurchaseSide() {
-		Label quantityLabel = new Label("Quantity");
-		Button purchaseBtn = new Button("Purchase");
 
-		purchaseSide.setHgap(20);
-		purchaseSide.setVgap(10);
-		purchaseSide.add(quantityLabel, 1, 1);
-		purchaseSide.add(quantity, 2, 1);
-		initPurchaseBtn(purchaseBtn);
-		purchaseSide.add(purchaseBtn, 2, 2);
-		setBottom(purchaseSide);
+	
+
+	
+	
+	 public class StockView extends TableView<StockItem>{
 		
-	}
-	
-	private void initPurchaseBtn(Button btn) {
-		btn.setOnAction(e -> {
-			new Alert(AlertType.CONFIRMATION).show();
-			try {
-				leftSide.purchaseItem(quantity.getQuantity());
-			} catch (ViewException e1) {
-				e1.showAlert();
-			}
-			quantity.reset();
-		});
-	}
-	
-	
-	 private class StockView extends VBox{
-		
-		private TableView<StockItem> stockTable = new TableView<>();
+		//private TableView<StockItem> stockTable = new TableView<>();
 		private ObservableList<StockItem> items = FXCollections.observableArrayList(StockController.getItems());
 		
 		public StockView() {
 			initTable();
-			setPadding(new Insets(10));
-			getChildren().add(stockTable);
 		}
 		
-		public void purchaseItem(int quantity) throws ViewException {
-			StockItem i = stockTable.getSelectionModel().getSelectedItem();
-			if(i != null )
-				StockController.purchaseStock(i, quantity);
+		public void purchaseItem() throws ViewException {
+			StockItem i = getSelectionModel().getSelectedItem();
+			if(i != null ) {
+				new PurchasePopUp(i, this).start(new Stage());
+				refresh();
+			}
 			else {
 				throw new ViewException("Please select an item", AlertType.ERROR);
 			}
-			refresh();
+		
 		}
 		
 		public void deleteItem() throws ViewException {
-			StockItem i = stockTable.getSelectionModel().getSelectedItem();
-			if(i != null )
-				StockController.delete(i);
+			StockItem i = getSelectionModel().getSelectedItem();
+			if(i != null ) {
+				Alert alert = new Alert(AlertType.CONFIRMATION, "Do you want to delete that item?",
+						ButtonType.YES, ButtonType.NO);
+				Optional<ButtonType> butons = alert.showAndWait();
+				if(butons.get() == ButtonType.YES)
+					StockController.delete(i);
+			}
 			else {
 				throw new ViewException("Please select an item", AlertType.ERROR);
 			}
@@ -104,8 +97,10 @@ public class PurchaseStock extends BorderPane {
 			addColumn("itemName", "Name");
 			addColumn("stockQuantity", "Quantity");
 			addColumn("purchasingPrice", "Purchasing Price");
-			stockTable.setPrefWidth(300);
-			stockTable.setItems(items);
+			setPrefWidth(600);
+			setPrefHeight(580);
+			setColumnResizePolicy(TableView.CONSTRAINED_RESIZE_POLICY);
+			setItems(items);
 			
 		}
 		
@@ -113,13 +108,18 @@ public class PurchaseStock extends BorderPane {
 			items.setAll(StockController.getItems());
 		}
 		
-		private void addColumn( String attributeName, String columnName) {
+		private void addColumn(String attributeName, String columnName) {
 			TableColumn<StockItem, String> column = new TableColumn<>(columnName);
 			column.setCellValueFactory(
 					new PropertyValueFactory<>(attributeName));
-			stockTable.getColumns().add(column);
+			getColumns().add(column);
 		}
 		
 		
-	}	
+	}
+
+
+
+
+	
 }
